requirements.txt
tensorflow-cpu==2.10.1
numpy==1.22.4
scipy
pandas
nltk
certifi
